fix(cloudwatch): math expressions incorrectly warn about metricsinsights variable names (#23316)
It is intended that all metric identifiers referenced in a MathExpression are included in the usingMetrics map and we will raise warnings if the customer does not follow this contract. 
However for Metricsinsights queries, we can refer directly to metrics attribute values inside the query. Therefore we do not raise warnings for Metricsinsights queries for not referencing metrics

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*