feat(s3): use Bucket Policy for Server Access Logging grant (under feature flag)
Using ACLs to grant access to buckets is no longer recommended. In fact,
it doesn't work if Object Ownership is set to be enforced for the
bucket. According to the service documentation for [enabling server
access logging][1], it is now preferred to use a bucket policy to grant
permission to deliver logs to a bucket.

Changing the default would result in changes to deployed resources, so
the new behavior is added behind a feature flag.

[1]: https://docs.aws.amazon.com/AmazonS3/latest/userguide/enable-server-access-logging.html