feat(assertions): improve printing of match failures (#23453)
Debugging match failures with the `assertions` library used to be a bit frustrating. The library used to guess the single closest matching resource and print it fully, with a list of match failures underneath.

This has two problems:

- Matching the failures up to the data structure requires a lot of scanning back and forth between different screen lines.
- The library could incorrectly guess at the closest match, giving you useless information.

This change is a first (and by no means the last) stab at improving the situation:

- Mismatch errors are printed in-line with the data structure, making it very clear what they are referencing (removing the need for scanning).
- Fields in complex data structures that aren't involved in the mismatch at all are collapsed. For example, a complex subobject that matches be printed as `{ ... }` so it doesn't take up too much vertical space.
- Because we now have a lot of vertical space left, we can print the N closest matches, lowering the chance that we guess wrong and leave you without information to correct the test.

The last point can probably be improved more in the future, by coming up with a more accurate metric for "closest match" than "failure count", but this will do for now.

----
*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*