refactor(aws-ec2): always resolve service name when looking up supported AZs
To lookup supported AZs it's required a resolved string for the service
name, however, in some cases the service name uses tokens that must be
resolved before the lookup. This problem is already solved for instances
of `InterfaceVpcEndpointAwsService` where the tokens are resolved only
for instances of this class, however, other instances from customers may
also contain unresolved tokens, and currently the only way to prevent a
failure is extending the `InterfaceVpcEndpointAwsService` class and
override everything (which is really a false inheritance).

A better approach may be to always attempt to resolve the service name
if it's an unresolved token because, and if for some reason the token
can't be resolved it's ok since it would fail anyways later when doing
the lookup.