fix(servicecatalog): make assetBuckets a required property
The `IProduct#assetBuckets` property was optional, but `ProductBase` did
not provide an implementation for it, causing the TypeScript type
checker to deem it opted out of, which the jsii compiler does not
currently check against. This results in code that fails to build in C#.

Additionally, the property was marked as read-only, making it impossible
to turn an `undefined` value to a valid array, which at least one unit
test would have relied on doing (or would have relied on a strong
assumption that a particular subclass would never return `undefined`,
which is risky).

Finally, the `CloudFormationProduct` class was providing a _required_
property implementation, which was perceived as a new, independent
property by the compiler, instead of as an implementation to the
interface's property.

To address these problems, turn the property required, and created
implementations where there were none.