fix(servicecatalog): make assetBuckets a required property (#23507)
The `IProduct#assetBuckets` property was optional, but `ProductBase` did not provide an implementation for it, causing the TypeScript type checker to deem it opted out of, which the jsii compiler does not currently check against. This results in code that fails to build in C#.

Additionally, the property was marked as read-only, making it impossible to turn an `undefined` value to a valid array, which at least one unit test would have relied on doing (or would have relied on a strong assumption that a particular subclass would never return `undefined`, which is risky).

Finally, the `CloudFormationProduct` class was providing a _required_ property implementation, which was perceived as a new, independent property by the compiler, instead of as an implementation to the interface's property.

To address these problems, turn the property required, and created implementations where there were none.


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Construct Runtime Dependencies:

* [ ] This PR adds new construct runtime dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-construct-runtime-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*