fix(events): cross stack rules require concrete environment
When a rule and a rule target are in different environments (account
and/or region) we have to perform some extra setup to get the cross
environment stuff working. Previously if one of the environments was not
defined then we assumed that we were working in a cross environment
fashion and we threw an error message. It is probably a much more common
scenario for both the stacks to be deployed to the same environment if
one of the environments is not defined.

This PR updates the logic to make that assumption and to provide a
warning to the user that if they _are_ working in a cross environment
setup they need to provide the environment.

fix #18405