feat(route53-patterns): use `Certificate` as the default certificate (under feature flag)
The `HttpsRedirect` construct creates a certificate using the
`DnsValidatedCertificate` if a certificate is not provided by the user.
As part of [deprecating]() the `DnsValidatedCertificate` we need to
replace all instances of `DnsValidatedCertificate` with `Certificate`.

I have placed this behind a feature flag since I think it should be
opt-in behavior for existing users, _but_ it is not a breaking change
because of the way CloudFormation works. If the flag is enabled on an
existing `HttpsRedirect` CloudFormation will:

1. Create a new certificate using `Certificate`
2. Update CloudFront to use the newly created `Certificate`
3. Delete the old certificate created with the `DnsValidatedCertificate`

I have tested this scenario using the two newly created integration
tests.