feat(pipelines): Expose stack output namespaces in custom `pipelines.Step`s (#23110)
Implements https://github.com/aws/aws-cdk/issues/23000 as per request from https://github.com/aws/aws-cdk/issues/23000#issuecomment-1324379670.

In order for custom steps to include `CfnOutput` in their action configurations, there needed to be a way to access and/or generate the output variable namespaces. `ShellStep` already had this capability.

This change generalizes the `StackOutputReference` usages by letting implementors of `Step` define the `StackOutputReference`s their step consumes, instead of hardwiring it to `ShellStep`.

To actually consume the references, the `ICodePipelineActionFactory` provides a `StackOutputsMap` that exposes a method to 
 render `StackOutputReference`s into their assigned CodePipeline variable names.