fix(aws-custom-resource): switch off `installLatestAwsSdk` by default
The `AwsCustomResource` reaches out to the internet to install the
latest AWS SDK by default. This will make it fail if it is being bound
to a VPC that doesn't have internet connectivity, or in
regions/partitions that are not able to freely connect to `npmjs.com`.

This was a poorly chosen default from the time we didn't know any
better, but we do know right now. Switch the behavior off by default
(under feature flag), and explicitly disable it for all
`AwsCustomResource`s the L2 library uses. Lambda advertises 2.1055.0
of the SDK everywhere, and I checked to make sure that all APIs we
use are part of that SDK version, so we don't need any newer version.

That version is a year old (!) so this is not the end of the story,
but it's at least an improvement over what we currently have.

Fixes #23113.