chore(cli): move integ tests to separate package
Clean up the CLI integration tests a bit, by moving
them to a separate package. This has the following
benefits:

- More clearly isolates the tests and keeps them separate from the CLI
  package proper.
- Allows normal specification of dependencies of the integration tests
  themselves, via a proper `package.json` and Yarn lockfile.
- Makes it more clear what reusable files are part of the CLI, vs
  what reusable files are part of the integration tests.
- Makes it more clear/easier to download an old copy of the tests.

Also makes the following improvement:

- Publishes the test artifacts to CodeArtifact, so that we don't need to
  do crazy shell hijacking shenanigans in order to fake
  NPM/Maven/NuGet/PyPI repositories from a set of artifacts.

This is not the final form of our testing infrastructure. Right now,
a fresh CodeArtifact repo is created for every test with all packages
in it. In an actual pipeline, we would be better off publishing to a
CodeArtifact repo once and all tests pulling from that.

That will be forthcoming in the future.