fix(lambda): inconsistent Version hash when LayerVersions from tokens
If `LayerVersions` are referenced using tokens
(`LayerVersion.fromLayerVersionArn(this, 'Layer', /* some deploy-time
value */`) then the version hash would incorrectly use the string
representation of the tokenized ARN and be different on every
deployment, incorrectly trying to create a new `Version` object on every
deployment.

Resolve the ARN if we detect this.

However, this will not be complete: we now have the problem that
a new Version will not be created if it were necessary, since CDK
cannot read the deploy-time value of the ARN and cannot mix it into
the Version LogicalID if necessary.

To fix that, add a:

```ts
fn.invalidateVersionBasedOn(...);
```

Function to help invalidate the version using outside information.