Adds gen script for aws-cdk-lib
Adds a `gen.ts` file in `scripts` for aws-cdk-lib which is meant to run
as the first step during build in order to generate L1 construct source
files. Added support for `*` as a scope in `cfn2ts` in order to enable
generating all desired L1s without declaring scopes in package.json.