chore(prlint): require consistent scope (without `aws-`) (#23672)
The AWS CDK uses conventional commits which encourage a scope. In nearly
all cases, contributors omit the `aws-` prefix; though there are some
situations where they're included (such as #23552, #23225, and others).
Interestingly, it seems that most often the `aws-` prefix is used for
chore/doc commits that don't end up in changelogs anyway.

This actually results in inconsistencies for users in the changelog.
Because the changelog sorts entries alphabetically by scope, changes
that were contributed with `aws-s3` as the scope are listed at the top
of the changelog whereas changes that just used `s3` are sorted further
down. This means it's harder for users to review one (or I suppose 2
with Feature/Fix being separate sections) spots in the changelog to
identify specific modules they care about.

It's also possible that rather than potentially annoying contributors
who have to edit their titles, this should be automatically fixed up in
the changelog generation. The current behavior may also accidentally
encourage users to scan the whole changelog rather than just a few bits
of it.

So it's very understandable if the potential inconvenience from this
change it's worth the minor formatting consistency gain.


----

### All Submissions:

* [X] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Construct Runtime Dependencies:

* [ ] This PR adds new construct runtime dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-construct-runtime-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*