fix(pipelines): cross-stack step dependencies have wrong name (#23594)
We used to immediately recurse into the dependencies of a Step, and add a new Node for them to the current Graph if they didn't exist yet.

This was originally written like this to support `inputs` for the `ShellSteps`, which would automatically add in any necessary Nodes into the same Graph as necessary without explicitly adding them.

However, it would also have the effect that Steps which **would have** been added to the Graph would be parented incorrectly (they would be eagerly added to the wrong parent).

Do this in two stages now: first try to resolve all step dependencies between things that are added to the graph, then create new Nodes for missing Step dependencies later.

Fixes #21843.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*