fix(aws-codeguruprofiler): correctly configure imported profiling group
This resource missed the common pattern of configuring the resource
environment from the arn, which results in the region being that of the
stack and not the profiling group. This means that integrations that
look at this (like the lambda profiling integration) will configure the
wrong region for the resource.

Additionally, exposes `profilingGroupArn` for consumption in other resources.