fix(integ-runner): cleanup tmp snapshot before running test
The integ-runner obtains information on how to run a test from the
`integ.json` manifest file in the cloud assembly. In order to get this
information for a new version of the test it first synthesizes the new
cloud assembly _only for the purpose of loading the integ manifest_. It
will then run the actual test which synthesizes again into the same
cloud assembly directory.

We need to cleanup that first temporary cloud assembly _before_ running
the actual test to make sure we remove any leftover files. Here is one
example of what might happen.

1. Do the temporary synthesis and create the cloud assembly at
   `cdk-integ.out.my-test.js/`. This assembly has an asset with some
   hash `asset.abcdefg`.
2. Do the synthesis while executing the test. The cloud assembly at
   `cdk-integ.out.my-test.js` is updated. This time the asset hash is
   different `asset.123456`. Now we have two assets in the cloud
   assembly and one `asset.abcdefg` is not being used!
3. Test completes and copies the temporary snapshot to the final
   snapshot directory, including the old asset

I've also added back the `lambda-nodejs/integ.dependencies-pnpm.ts`
integration test that was removed in #23728. With this fix the test will
no longer check in the asset file. I also removed the Trigger from the
test since that is what introduced the always changing diff.