feat(elbv2): add metrics to INetworkLoadBalancer and IApplicationLoadBalancer
By moving the metrics methods to the `INetworkLoadBalancer` and
`IApplicationLoadBalancer` interfaces it allows to create these metrics also for
LBs that are imported via the `fromXXX` methods.

To create the metrics for LBs requires only the full name of the LB. This
attribute is available at the constructs returned by the `fromXXX` methods.

To solve this problem I did:

- Introduce a new interface for each LB type: `INetworkLoadBalancerMetrics`,
`IApplicationLoadBalancerMetrics`
- Create a concrete implementation for the new interfaces (1 for each):
`NetworkLoadBalancerMetrics` and `ApplicationLoadBalancerMetrics`
- Make each concrete implementation of each Load Balancer to also provide a
`metrics` field. The concrete implementations of the load balancers are:
`ImportedApplicationLoadBalancer`, `LookedUpApplicationLoadBalancer`,
`ApplicationLoadBalancer` (and the same for the NLB classes).

I chose to create a new interface because code can be reused across the 3
concrete implementations of each Load Balancer. I deprecated the `metricXXX()`
methods of each load balancer because I think it is cleaner to access metrics
through the new `metrics` attribute/interface.