chore: remove strict check on command arguments
As part of #23907 I updated the dependency on `@types/yargs` to be in
the `15` major version to reflect the version we use in the rest of the
repo. Version `15` did not have the `strictOptions` yargs method so I
attempted to enable `strict()` and disable `strictCommands` but it looks
like that doesn't work. Removing strict altogether.