feat: remodel calls ubergen
The remodel tool now calls some functions from ubergen in order to
create the aws-cdk-lib package in a way that is consistent with how it
is built today. Some differences from ubergen are required since it
assumes that codegen has already taken place, and that L1s need to be
extracted from experimental modules, which is not the case when working
on aws-cdk-lib directly.

Additionally makes a bunch of the functions in ubergen exported and
changes them to rely on a passed configuration value instead of global
variables, so that we can correctly call it in the directories that we
want.