fix(codedeploy): cross-region referenced groups use wrong config
DeploymentGroups grant IAM permissions to their DeploymentConfigs, by
their ARN.

When using a predefined DeploymentConfig (like `ECS_ALL_AT_ONCE`)
however, we used to use `Aws.ACCOUNT_ID, Aws.REGION` to build the
ARN for the DeploymentConfig. This would be incorrect if the
DeploymentGroup is referenced in a different region (by using
`DeploymentGroup.fromDeploymentGroupArn()`): the `Aws.REGION` token
would resolve to the region of the *referencing Stack*, instead of
the region of the *referenced DeploymentGroup*.

Make all predefined DeploymentConfigs implement a hidden interface
that allows the DeploymentGroup to specialize the Config to its
own account and region. This behavior is not relevant for user-created
deployment configs: those will still be region-bound, so customers
won't need access to this interface.