feat(elbv2): add metrics to INetworkTargetGroup and IApplicationTargetGroup
By moving the metrics methods to the `INetworkTargetGroup` and
`IApplicationTargetGroup` interfaces it allows to create these metrics also for
Target Groups that are imported via the `fromTargetGroupAttributes()` method.

To create the metrics for Target Groups requires (1) the full name of the Target
Group and (2) the full name of the Load Balancer.

For (1): it is readily available given that all imported Target Groups need to
provide its ARN.
For (2), it is an optional value, so the `.metrics` parameter will throw an
error if it was not provided.

To solve this problem I did:

- Introduce a new interface for each TG type: `INetworkTargetGroupMetrics`,
`IApplicationTargetGroupMetrics`
- Create a concrete implementation for the new interfaces (1 for each):
`NetworkTargetGroupMetrics` and `ApplicationTargetGroupMetrics`
- Make each concrete implementation of each Load Balancer to also provide a
`metrics` field. The concrete implementations of the load balancers are:
`ImportedApplicationTargetGroup`, and `ApplicationLoadBalancer` (and the same
for the NLB classes).

I chose to create a new interface because code can be reused across the 3
concrete implementations of each Load Balancer. I deprecated the `metricXXX()`
methods of each load balancer because I think it is cleaner to access metrics
through the new `metrics` attribute/interface.

There is a small **gotcha** here because the parameter of the
`fromTargetGroupAttributes()` method that refers to the LB is:
`loadBalancerArns`, which has its documentation as:

> A Token representing the list of ARNs for the load balancer routing to this
target group

I'm not treating this parameter as a collection of ARNs, but as a single AR.
Also, I'm not treating it only as a token, but hardcoded ARNs can also be
supplied, which "sort of" violates its interface. This attribute is weird though
because Target Groups cannot have multiple Load Balancers as of today, although
its documentation doesn't clearly express that is the case.

fix: #10850

----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing
guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Construct Runtime Dependencies:

* [ ] This PR adds new construct runtime dependencies following the process
described
[here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-construct-runtime-dependencies)

### New Features

* [x] Have you added the new feature to an [integration
test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [x] Did you use `yarn integ` to deploy the infrastructure and generate the
snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache-2.0 license*